+++ src/hihapi/Controllers/Finance/FinanceControlCentersController.cs
@@ -86,6 +86,7 @@ namespace hihapi.Controllers
         {
             if (!ModelState.IsValid)
             {
+#if DEBUG
                 foreach (var value in ModelState.Values)
                 {
                     foreach(var err in value.Errors) 
@@ -93,6 +94,7 @@ namespace hihapi.Controllers
                         System.Diagnostics.Debug.WriteLine(err.Exception?.Message);
                     }
                 }
+#endif
 
                 return BadRequest();
             }
@@ -119,7 +121,7 @@ namespace hihapi.Controllers
                 throw new UnauthorizedAccessException();        
             }
 
-            if (!controlCenter.IsValid())
+            if (!controlCenter.IsValid(this._context))
                 return BadRequest();
 
             _context.FinanceControlCenter.Add(controlCenter);
@@ -133,6 +135,7 @@ namespace hihapi.Controllers
         {
             if (!ModelState.IsValid)
             {
+#if DEBUG
                 foreach (var value in ModelState.Values)
                 {
                     foreach(var err in value.Errors) 
@@ -140,6 +143,7 @@ namespace hihapi.Controllers
                         System.Diagnostics.Debug.WriteLine(err.Exception?.Message);
                     }
                 }
+#endif
 
                 return BadRequest();
             }
@@ -170,7 +174,7 @@ namespace hihapi.Controllers
                 throw new UnauthorizedAccessException();        
             }
 
-            if (!update.IsValid())
+            if (!update.IsValid(this._context))
                 return BadRequest();
 
             _context.Entry(update).State = EntityState.Modified;
@@ -201,6 +205,9 @@ namespace hihapi.Controllers
             {
                 return NotFound();
             }
+
+            if (!cc.IsDeleteAllowed(this._context))
+                return BadRequest();
 
             _context.FinanceControlCenter.Remove(cc);
             await _context.SaveChangesAsync();
