+++ src/hihapi/Controllers/Finance/FinanceOrdersController.cs
@@ -87,6 +87,7 @@ namespace hihapi.Controllers
         {
             if (!ModelState.IsValid)
             {
+#if DEBUG
                 foreach (var value in ModelState.Values)
                 {
                     foreach (var err in value.Errors)
@@ -94,12 +95,13 @@ namespace hihapi.Controllers
                         System.Diagnostics.Debug.WriteLine(err.Exception?.Message);
                     }
                 }
+#endif
 
                 return BadRequest();
             }
 
             // Check
-            if (!order.IsValid())
+            if (!order.IsValid(this._context))
             {
                 return BadRequest();
             }
@@ -126,7 +128,7 @@ namespace hihapi.Controllers
                 throw new UnauthorizedAccessException();
             }
 
-            if (!order.IsValid())
+            if (!order.IsValid(this._context))
                 return BadRequest();
 
             _context.FinanceOrder.Add(order);
@@ -145,6 +147,7 @@ namespace hihapi.Controllers
         {
             if (!ModelState.IsValid)
             {
+#if DEBUG
                 foreach (var value in ModelState.Values)
                 {
                     foreach (var err in value.Errors)
@@ -152,6 +155,7 @@ namespace hihapi.Controllers
                         System.Diagnostics.Debug.WriteLine(err.Exception?.Message);
                     }
                 }
+#endif
 
                 return BadRequest();
             }
@@ -182,7 +186,7 @@ namespace hihapi.Controllers
                 throw new UnauthorizedAccessException();
             }
 
-            if (!update.IsValid())
+            if (!update.IsValid(this._context))
                 return BadRequest();
 
             _context.Entry(update).State = EntityState.Modified;
@@ -213,6 +217,9 @@ namespace hihapi.Controllers
             {
                 return NotFound();
             }
+
+            if (!cc.IsDeleteAllowed(this._context))
+                return BadRequest();
 
             _context.FinanceOrder.Remove(cc);
             await _context.SaveChangesAsync();
