+++ src/hihapi/Controllers/Finance/FinanceTransactionTypesController.cs
@@ -61,6 +61,7 @@ namespace hihapi.Controllers
         {
             if (!ModelState.IsValid)
             {
+#if DEBUG
                 foreach (var value in ModelState.Values)
                 {
                     foreach (var err in value.Errors)
@@ -68,12 +69,13 @@ namespace hihapi.Controllers
                         System.Diagnostics.Debug.WriteLine(err.Exception?.Message);
                     }
                 }
+#endif
 
                 return BadRequest();
             }
 
             // Check
-            if (!ctgy.IsValid() || !ctgy.HomeID.HasValue)
+            if (!ctgy.IsValid(this._context) || !ctgy.HomeID.HasValue)
             {
                 return BadRequest();
             }
@@ -100,7 +102,7 @@ namespace hihapi.Controllers
                 throw new UnauthorizedAccessException();
             }
 
-            if (!ctgy.IsValid())
+            if (!ctgy.IsValid(this._context))
                 return BadRequest();
 
             _context.FinTransactionType.Add(ctgy);
@@ -114,6 +116,7 @@ namespace hihapi.Controllers
         {
             if (!ModelState.IsValid)
             {
+#if DEBUG
                 foreach (var value in ModelState.Values)
                 {
                     foreach (var err in value.Errors)
@@ -121,6 +124,7 @@ namespace hihapi.Controllers
                         System.Diagnostics.Debug.WriteLine(err.Exception?.Message);
                     }
                 }
+#endif
 
                 return BadRequest();
             }
@@ -151,7 +155,7 @@ namespace hihapi.Controllers
                 throw new UnauthorizedAccessException();
             }
 
-            if (!update.IsValid())
+            if (!update.IsValid(this._context))
                 return BadRequest();
 
             _context.Entry(update).State = EntityState.Modified;
@@ -182,6 +186,9 @@ namespace hihapi.Controllers
             {
                 return NotFound();
             }
+
+            if (!cc.IsDeleteAllowed(this._context))
+                return BadRequest();
 
             _context.FinTransactionType.Remove(cc);
             await _context.SaveChangesAsync();
